demo/requirements.txt
Django==1.5.1
Unipath==1.0
djangorestframework==2.2.7
